fix: RTA_TransactionRecordBase::Amount is uint64_t

+++ src/supernode/supernode_common_struct.h
@@ -71,7 +71,7 @@ namespace supernode {
 
 	// ------------------------------------------
 	struct RTA_TransactionRecordBase : public SubNetData {
-        unsigned Amount;
+        uint64_t Amount;
         string POSAddress;
         string POSSaleDetails;// empty in wallet call
         uint64_t BlockNum;// empty in pos call
